requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
python-environ==0.4.54
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
simplejson==3.19.2
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
